fix(phdconfig): serialize integer config values with %ld, not %lu
Address the automated review on PR #9: WriteVal serialized a `long` config
value with `%lu` and an (unsigned long) cast, which would wrap a negative value
to a huge number that wouldn't round-trip back through the integer parser. The
value is a signed `long`, so `%ld` is both the correct format and a clean fix
for the cppcheck arg-type finding (no cast needed).

+++ src/phdconfig.cpp
@@ -877,7 +877,7 @@ static void WriteVal(wxTextOutputStream& os, wxConfigBase *cfg, const wxString& 
     {
         long val = 0;
         (void) cfg->Read(key, &val);
-        sval = wxString::Format("%lu", (unsigned long) val);
+        sval = wxString::Format("%ld", val);
         break;
     }
     case wxConfigBase::Type_Float:
